requirements.txt
streamlit==1.33.0
pandas
reportlab
openpyxl
matplotlib
seaborn
mysql.connector
sqlalchemy
mysqlclient
logging